refactor: Fix broken imports after rename
